terminal_input, clr_screen

+++ Libraries/API-LIB/Header/triangle.h
@@ -10,6 +10,6 @@
 
 #include "includes.h"
 
-void draw_triangle(int x1,int y1,int x2,int y2,int xtop,int ytop, uint8_t color);
+void draw_triangle(int x1,int y1,int x2,int y2,int xtop,int ytop, int color);
 
 #endif /* API_LIB_HEADER_TRIANGLE_H_ */
